split publisher service methods into topic-oriented files
